Ensure Page is a 1 not a 0 index, starts creating base test class for controllers to ensure we are testing everything, moves some logic to a shared class

+++ src/Umbraco.RestApi/Models/ContentPagedListRepresentation.cs
@@ -5,7 +5,7 @@ namespace Umbraco.RestApi.Models
 {
     public class ContentPagedListRepresentation : PagedListRepresentation<ContentRepresentation>
     {
-        public ContentPagedListRepresentation(IList<ContentRepresentation> res, int totalResults, int totalPages, int page, int pageSize, Link uriTemplate, object uriTemplateSubstitutionParams) :
+        public ContentPagedListRepresentation(IList<ContentRepresentation> res, long totalResults, long totalPages, long page, int pageSize, Link uriTemplate, object uriTemplateSubstitutionParams) :
             base(res, totalResults, totalPages, page, pageSize, uriTemplate, uriTemplateSubstitutionParams)
         {
         }
