+++ src/Umbraco.RestApi/Models/MediaPagedListRepresentation.cs
@@ -5,7 +5,7 @@ namespace Umbraco.RestApi.Models
 {
     public class MediaPagedListRepresentation : PagedListRepresentation<MediaRepresentation>
     {
-        public MediaPagedListRepresentation(IList<MediaRepresentation> res, int totalResults, int totalPages, int page, int pageSize, Link uriTemplate, object uriTemplateSubstitutionParams) :
+        public MediaPagedListRepresentation(IList<MediaRepresentation> res, long totalResults, long totalPages, long page, int pageSize, Link uriTemplate, object uriTemplateSubstitutionParams) :
             base(res, totalResults, totalPages, page, pageSize, uriTemplate, uriTemplateSubstitutionParams)
         {
         }
