+++ src/Umbraco.RestApi/Models/MemberPagedListRepresentation.cs
@@ -5,7 +5,7 @@ namespace Umbraco.RestApi.Models
 {
     public class MemberPagedListRepresentation : PagedListRepresentation<MemberRepresentation>
     {
-        public MemberPagedListRepresentation(IList<MemberRepresentation> res, int totalResults, int totalPages, int page, int pageSize, Link uriTemplate, object uriTemplateSubstitutionParams) :
+        public MemberPagedListRepresentation(IList<MemberRepresentation> res, long totalResults, long totalPages, long page, int pageSize, Link uriTemplate, object uriTemplateSubstitutionParams) :
             base(res, totalResults, totalPages, page, pageSize, uriTemplate, uriTemplateSubstitutionParams)
         {
         }
