+++ src/Umbraco.RestApi/Models/PublishedContentPagedListRepresentation.cs
@@ -5,7 +5,7 @@ namespace Umbraco.RestApi.Models
 {
     public class PublishedContentPagedListRepresentation : PagedListRepresentation<PublishedContentRepresentation>
     {
-        public PublishedContentPagedListRepresentation(IList<PublishedContentRepresentation> res, int totalResults, int totalPages, int page, int pageSize, Link uriTemplate, object uriTemplateSubstitutionParams) : 
+        public PublishedContentPagedListRepresentation(IList<PublishedContentRepresentation> res, long totalResults, long totalPages, long page, int pageSize, Link uriTemplate, object uriTemplateSubstitutionParams) : 
             base(res, totalResults, totalPages, page, pageSize, uriTemplate, uriTemplateSubstitutionParams)
         {
         }
